put all the builtins in a main, moved pid reset, some malloc prots, small stuff

+++ src/minishell_main.c
@@ -6,7 +6,7 @@
 /*   By: bvercaem <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/23 14:00:30 by bvercaem          #+#    #+#             */
-/*   Updated: 2023/10/26 16:04:49 by bvercaem         ###   ########.fr       */
+/*   Updated: 2023/10/27 15:53:28 by bvercaem         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,36 +14,74 @@
 
 pid_t	pid;
 
-void	ms_exit(int exit_value)
+// if exit_value == -5, flushes all then just returns (not in use i think)
+void	ms_exit(t_shellshock *data, int exit_value)
 {
-	write(STDOUT_FILENO, "exit\n", 5);
 	rl_clear_history();
+	ft_clear_ds(data->env);
+	if (exit_value == -5)
+		return ;
+	write(STDOUT_FILENO, "exit\n", 5);
 	exit(exit_value);
 }
 
-// should free line
-int	ms_process_line(char *line)
+// returns 1 if something was executed
+int	ms_parse_builtins(t_shellshock *data, char *line)
 {
 	if (!ft_strncmp(line, "exit", 5))
 	{
 		free(line);
-		ms_exit(0);
+		ms_exit(data, 0);
 	}
-	if (!ft_strncmp(line, "sleep", 6))
+	if (!ft_strncmp(line, "cd ", 3))
+		ms_cd(data, line + 3);
+	else if (!ft_strncmp(line, "pwd", 4))
+		ms_pwd(data);
+	else if (!ft_strncmp(line, "export", 7))
+		ms_export(data, NULL);
+	else if (!ft_strncmp(line, "export ", 7))
+		ms_export(data, ft_split(line + 7, ' '));
+	else if (!ft_strncmp(line, "unset ", 6))
+		ms_unset(data, ft_split(line + 6, ' '));
+	else if (!ft_strncmp(line, "env", 4))
+		ms_env();
+	else if (!ft_strncmp(line, "echo ", 5))
+		ms_echo(ft_split(line + 5, ' '));
+	else if (!ft_strncmp(line, "echo", 5))
+		ms_echo(NULL);
+	else
+		return (0);
+	return (1);
+}
+
+static void	ms_sleep(char *line)
+{
+// TEST FUNCTION
+	extern char	**environ;
+	char		*argv[2];
+
+	pid = fork ();
+	if (!pid)
 	{
-		pid = fork ();
-		if (!pid)
-		{
-			free (line);
-			rl_clear_history();
-			execve ("sleep", NULL, NULL);
-			exit (1);
-		}
-		if (pid)
-			waitpid(pid, NULL, 0);
-		pid = 0;
+		free (line);
+		rl_clear_history();
+		argv[0] = "./sleep";
+		argv[1] = NULL;
+		execve ("sleep", argv, environ);
+		exit (1);
 	}
-// do more stuff :)
+	if (pid)
+		waitpid(pid, NULL, 0);
+}
+
+// should free line
+int	ms_process_line(t_shellshock *data, char *line)
+{
+	if (ms_parse_builtins(data, line))
+		(void) line;
+	else if (!ft_strncmp(line, "sleep", 6))
+		ms_sleep(line);
+// do more stuff, just have some fun with it :)
 	if (*line)
 		add_history(line);
 	free(line);
@@ -52,7 +90,10 @@ int	ms_process_line(char *line)
 
 void	ms_new_prompt(int sig)
 {
-	(void) sig;
+	if (sig != SIGINT)
+		return ;
+	if (!pid)
+		write(STDOUT_FILENO, "^C", 2);
 	write(STDOUT_FILENO, "\n", 1);
 	if (pid)
 		return ;
@@ -63,7 +104,8 @@ void	ms_new_prompt(int sig)
 
 void	ms_kill_pid(int sig)
 {
-	(void) sig;
+	if (sig != SIGQUIT)
+		return ;
 	if (pid)
 	{
 		kill (pid, SIGQUIT);
@@ -79,8 +121,11 @@ int	main(void)
 	struct sigaction	act_sigint;
 	struct sigaction	act_sigquit;
 
-	pid = 0;
 	rl_catch_signals = 0;
+	data.env_excess = 0;
+	if (ms_envcpy(&data))
+		return (1);
+// malloc fail
 	ft_bzero(&act_sigint, sizeof(struct sigaction));
 	ft_bzero(&act_sigquit, sizeof(struct sigaction));
 	act_sigint.__sigaction_u.__sa_handler = ms_new_prompt;
@@ -90,8 +135,9 @@ int	main(void)
 	line = readline(MS_PROMPT);
 	while (line)
 	{
-		ms_process_line(line);
+		pid = 0;
+		ms_process_line(&data, line);
 		line = readline(MS_PROMPT);
 	}
-	ms_exit(0);
+	ms_exit(&data, 0);
 }
